【core - build core - article】

+++ src/main/java/com/shnnny/notBlog/model/po/Article.java
@@ -29,12 +29,12 @@ public class Article implements Serializable {
     /**
      * 内容生成时的GMT unix时间戳
      */
-    private Integer created;
+    private Long created;
 
     /**
      * 内容更改时的GMT unix时间戳
      */
-    private Integer modified;
+    private Long modified;
 
     /**
      * 内容所属用户id
@@ -121,20 +121,20 @@ public class Article implements Serializable {
     }
 
     @Column(length = 40)
-    public Integer getCreated() {
+    public Long getCreated() {
         return created;
     }
 
-    public void setCreated(Integer created) {
+    public void setCreated(Long created) {
         this.created = created;
     }
 
     @Column(length = 40)
-    public Integer getModified() {
+    public Long getModified() {
         return modified;
     }
 
-    public void setModified(Integer modified) {
+    public void setModified(Long modified) {
         this.modified = modified;
     }
